+++ scripts/run_archiso.sh
@@ -20,6 +20,7 @@ Usage:
     ${app_name} [options]
 
 Options:
+    -A [arch]       change architecture (defaults to the host architecture; supports: x86_64, i686, aarch64 and riscv64)
     -a              set accessibility support using brltty
     -b              set boot type to 'BIOS'
     -d              set image type to hard disk instead of optical disc
@@ -43,12 +44,32 @@ cleanup_working_dir() {
     fi
 }
 
-copy_ovmf_vars() {
-    if [[ ! -f '/usr/share/edk2/x64/OVMF_VARS.4m.fd' ]]; then
-        printf '[%s] ERROR: %s\n' "$app_name" "OVMF_VARS.4m.fd not found. Install edk2-ovmf." >&2
+check_architecture() {
+    if [[ "$boot_type" == 'bios' ]]; then
+        if [[ "$arch" != @('i686'|'x86_64') ]]; then
+            printf '[%s] Error: Unsupported architecture for the BIOS boot method: %s\n' "$app_name" "$arch" >&2
+            printf '[%s] Error: The BIOS boot method is supported on: %s\n' "$app_name" 'x86_64 i686' >&2
+            exit 1
+        fi
+    else
+        if ! [[ -v ovmf_code["$arch"] && -v ovmf_vars["$arch"] ]]; then
+            printf '[%s] Error: Unsupported architecture for the UEFI boot method: %s\n' "$app_name" "$arch" >&2
+            printf '[%s] Error: The UEFI boot method is supported on: %s\n' "$app_name" "${!ovmf_code[*]}" >&2
+            exit 1
+        elif ! [[ -f "${ovmf_code[$arch]}" && -f "${ovmf_vars[$arch]}" ]]; then
+            printf '[%s] ERROR: %s not found. Install OVMF for %s.\n' "$app_name" "${ovmf_vars[$arch]}" "$arch" >&2
+            exit 1
+        fi
+    fi
+
+    case "$arch" in
+        i686) qemu_command='qemu-system-i386' ;;
+        *) qemu_command="qemu-system-${arch}" ;;
+    esac
+    if ! command -v "$qemu_command" %>/dev/null; then
+        printf '[%s] ERROR: %s not found. Install QEMU for %s.\n' "$app_name" "$qemu_command" "$arch" >&2
         exit 1
     fi
-    cp -av -- '/usr/share/edk2/x64/OVMF_VARS.4m.fd' "${working_dir}/"
 }
 
 check_image() {
@@ -63,17 +84,39 @@ check_image() {
 }
 
 run_image() {
+    # Use KVM acceleration if possible
+    if [[ "$arch" == "$(uname -m)" || ( "$arch" == 'i686' && "$(uname -m)" == 'x86_64' ) ]]; then
+        qemu_options+=(-enable-kvm)
+    fi
+
+    # Set architecture-specific options
+    case "$arch" in
+        x86_64|i686)
+            qemu_options=(
+                -machine "type=q35,smm=on,usb=on,pcspk-audiodev=snd0"
+                -global ICH9-LPC.disable_s3=1
+                -vga virtio
+                "${qemu_options[@]}"
+            )
+            ;;
+        aarch64|riscv64)
+            qemu_options=(
+                -machine "type=virt,usb=on,acpi=on"
+                -device 'virtio-gpu-pci,id=video0'
+                "${qemu_options[@]}"
+            )
+            ;;
+    esac
+
     if [[ "$boot_type" == 'uefi' ]]; then
-        copy_ovmf_vars
+        cp -av -- "${ovmf_vars[$arch]}" "${working_dir}/"
         if [[ "${secure_boot}" == 'on' ]]; then
             printf '[%s] Using Secure Boot\n' "$app_name"
-            local ovmf_code='/usr/share/edk2/x64/OVMF_CODE.secboot.4m.fd'
-        else
-            local ovmf_code='/usr/share/edk2/x64/OVMF_CODE.4m.fd'
         fi
+
         qemu_options+=(
-            '-drive' "if=pflash,format=raw,unit=0,file=${ovmf_code},read-only=on"
-            '-drive' "if=pflash,format=raw,unit=1,file=${working_dir}/OVMF_VARS.4m.fd"
+            '-drive' "if=pflash,format=raw,unit=0,file=${ovmf_code[$arch]},read-only=on"
+            '-drive' "if=pflash,format=raw,unit=1,file=${working_dir}/${ovmf_vars[$arch]##*/}"
             '-global' "driver=cfi.pflash01,property=secure,value=${secure_boot}"
         )
     fi
@@ -92,29 +135,32 @@ run_image() {
         )
     fi
 
-    qemu-system-x86_64 \
+    "$qemu_command" \
         -boot order=d,menu=on,reboot-timeout=5000 \
         -m "size=3072,slots=0,maxmem=$((3072*1024*1024))" \
+        -cpu max \
+        -smp 4 \
         -k en-us \
         -name archiso,process=archiso_0 \
         -device virtio-scsi-pci,id=scsi0 \
-        -device "scsi-${mediatype%rom},bus=scsi0.0,drive=${mediatype}0" \
-        -drive "id=${mediatype}0,if=none,format=raw,media=${mediatype/hd/disk},read-only=on,file=${image}" \
         -display "${display}" \
-        -vga virtio \
         -audiodev pa,id=snd0 \
         -device ich9-intel-hda \
         -device hda-output,audiodev=snd0 \
         -device virtio-net-pci,romfile=,netdev=net0 -netdev user,id=net0,hostfwd=tcp::60022-:22 \
-        -machine type=q35,smm=on,accel=kvm,usb=on,pcspk-audiodev=snd0 \
-        -global ICH9-LPC.disable_s3=1 \
-        -enable-kvm \
+        -device qemu-xhci \
+        -device usb-kbd \
+        -device usb-mouse \
+        -device usb-tablet \
+        -device "scsi-${mediatype%rom},bus=scsi0.0,drive=${mediatype}0" \
+        -drive "id=${mediatype}0,if=none,format=raw,media=${mediatype/hd/disk},read-only=on,file=${image}" \
         "${qemu_options[@]}" \
         -serial stdio \
         -no-reboot
 }
 
 readonly app_name="${0##*/}"
+arch="$(uname -m)"
 image=''
 oddimage=''
 accessibility=''
@@ -122,13 +168,23 @@ boot_type='uefi'
 mediatype='cdrom'
 secure_boot='off'
 display='sdl'
+qemu_command=''
 qemu_options=()
 working_dir="$(mktemp -dt run_archiso.XXXXXXXXXX)"
+readonly -A ovmf_code=(['x86_64']='/usr/share/edk2/x64/OVMF_CODE.secboot.4m.fd'
+                       ['aarch64']='/usr/share/edk2/aarch64/QEMU_CODE.fd'
+                       ['riscv64']='/usr/share/edk2/riscv64/RISCV_VIRT_CODE.fd')
+readonly -A ovmf_vars=(['x86_64']='/usr/share/edk2/x64/OVMF_VARS.4m.fd'
+                       ['aarch64']='/usr/share/edk2/aarch64/QEMU_VARS.fd'
+                       ['riscv64']='/usr/share/edk2/riscv64/RISCV_VIRT_VARS.fd')
 trap cleanup_working_dir EXIT
 
 if (( ${#@} > 0 )); then
-    while getopts 'abc:dhi:suv' flag; do
+    while getopts 'A:abc:dhi:suv' flag; do
         case "$flag" in
+            A)
+                arch="${OPTARG,,}"
+                ;;
             a)
                 accessibility='on'
                 ;;
@@ -169,5 +225,6 @@ else
     exit 1
 fi
 
+check_architecture
 check_image
 run_image